separate Corpus and Novel from common -- for module teaching...
