+++ gender_novels/novel.py
@@ -1,0 +1,112 @@
+import string
+from pathlib import Path
+
+from gender_novels import common
+
+class Novel(common.FileLoaderMixin):
+    """ The Novel class loads and holds the full text and
+    metadata (author, title, publication date) of a novel
+
+    >>> from gender_novels import novel
+    >>> novel_metadata = {'author': 'Austen, Jane', 'title': 'Persuasion',
+    ...                   'corpus_name': 'sample_novels', 'date': '1818',
+    ...                   'filename': 'austen_persuasion.txt'}
+    >>> austen = novel.Novel(novel_metadata)
+    >>> type(austen.text)
+    <class 'str'>
+    >>> len(austen.text)
+    467018
+    """
+
+    def __init__(self, novel_metadata_dict):
+
+        if not hasattr(novel_metadata_dict, 'items'):
+            raise ValueError(
+                'novel_metadata_dict must be a dictionary or support .items()')
+
+        # Check that the essential attributes for the novel exists.
+        for key in ('author', 'date', 'title', 'corpus_name', 'filename'):
+            if key not in novel_metadata_dict:
+                raise ValueError(
+                    f'novel_metadata_dict must have an entry for "{key}"')
+
+        self.author = novel_metadata_dict['author']
+        self.data = novel_metadata_dict['date']
+        self.title = novel_metadata_dict['title']
+        self.corpus_name = novel_metadata_dict['corpus_name']
+        self.filename = novel_metadata_dict['filename']
+
+        # optional attributes
+        self.country_publication = novel_metadata_dict.get(
+            'country_publication', None)
+        self.author_gender = novel_metadata_dict.get('author_gender', None)
+        self.notes = novel_metadata_dict.get('notes', None)
+
+        if 'text' in novel_metadata_dict:
+            self.text = novel_metadata_dict['text']
+        else:
+            self.text = self._load_novel_text()
+
+    def _load_novel_text(self):
+        """Loads the text of a novel and removes
+        boilerplate at the beginning and end
+
+        Currently only supports boilerplate removal for
+        Project Gutenberg ebooks.
+
+        :rtype: str
+        """
+        file_path = Path('corpora', self.corpus_name, 'texts', self.filename)
+
+        try:
+            text = self.load_file(file_path)
+        except FileNotFoundError:
+            err = "Could not find the novel text file "
+            err += "at the expected location ({file_path})."
+            raise FileNotFoundError(err)
+
+        # Extract Project Gutenberg Boilerplate
+        if text.find('*** START OF THIS PROJECT GUTENBERG EBOOK') > -1:
+            end_intro_boilerplate = text.find(
+                '*** START OF THIS PROJECT GUTENBERG EBOOK')
+            # second set of *** indicates start
+            start_novel = text.find('***', end_intro_boilerplate + 5) + 3
+            end_novel = text.find('*** END OF THIS PROJECT GUTENBERG EBOOK')
+            text = text[start_novel:end_novel]
+
+        return text
+
+    def get_tokenized_text(self):
+        """
+        Tokenizes the text and returns it as a list of tokens
+
+        This is a very simple way of tokenizing the text. We will replace it soon with a
+        better implementation that uses either regex or nltk
+        E.g. this version doesn't handle dashes or contractions
+
+        >>> from gender_novels import novel
+        >>> novel_metadata = {'author': 'Austen, Jane', 'title': 'Persuasion', 'date': '1818',
+        ...                   'corpus_name': 'sample_novels', 'filename': 'austen_persuasion.txt',
+        ...                   'text': '?!All-kinds %$< of pun*ct(uatio)n {a}nd sp+ecial cha/rs'}
+        >>> austin = novel.Novel(novel_metadata)
+        >>> tokenized_text = austin.get_tokenized_text()
+        >>> tokenized_text
+        ['allkinds', 'of', 'punctuation', 'and', 'special', 'chars']
+
+        :rtype: list
+        """
+
+        # Excluded characters: !"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~
+        excluded_characters = set(string.punctuation)
+        cleaned_text = ''
+        for character in self.text:
+            if character not in excluded_characters:
+                cleaned_text += character
+
+        tokenized_text = cleaned_text.lower().split()
+        return tokenized_text
+
+
+if __name__ == '__main__':
+    from dh_testers.testRunner import main_test
+    main_test()
